delete box oauth files when not chosen

+++ hooks/post_gen_project.py
@@ -28,6 +28,12 @@ packages_to_install += ["awscli"]
 
 # {% if cookiecutter.dataset_storage.box %}
 packages_to_install += ["boxsdk"]
+
+# {% else %}
+# Remove the box oauth folder with code
+box_path = Path("{{ cookiecutter.module_name }}/utils/box_oauth")
+shutil.rmtree(box_path)
+
 # {% endif %} #
 
 
@@ -120,7 +126,13 @@ Path("pyproject.toml").write_text(pyproject_text.replace(r"\u0027", "'"))
 # remove everything except __init__.py so result is an empty package
 for generated_path in Path("{{ cookiecutter.module_name }}").iterdir():
     if generated_path.is_dir():
-        shutil.rmtree(generated_path)
+        # Leave box oauth utils alone if it is chosen for dataset storage
+        # {% if cookiecutter.dataset_storage.box %}
+        if generated_path.name == 'utils':
+            continue
+        # {% endif %}
+        shutil.rmtree(generated_path)   
+        
     elif generated_path.name != "__init__.py":
         generated_path.unlink()
     elif generated_path.name == "__init__.py":
